Refactored out the Constants and config to reuse.
